Update osp_processUnEdited.m
Add a fix to the processing of unedited data. Occasionally, the water removal routine will return a completely empty FID array. Add an intercept, and a loop that will change the number of components that are fit to the water signal and removed. As soon as a non-empty array is returned, processing continues.