Update spline function calculation (fit_makeSplineBasis)
Previously, we relied on MATLAB functions only available in the Curve Fitting Toolbox (spmak, fnval) to generate the cubic B-spline spline functions for baseline estimation.
These are now replaced with recursive iteration, eliminating the need for the toolbox altogether. In the process, the function has been cleaned up, removing the second derivatives that were never really used in the first place.