Bugfix in GUI job creation
Fix a bug in the function creating a job file from GUI inputs: Previously, only the first entry in a GUI list of multiple water reference datasets would be parsed correctly into the job file, while the rest of the entries would be erroneously replaced by the metabolite data. This is now corrected.