Add probabSpecReg
Added a multiplexed, probabilistic approach for aligning HERMES/HERCULES and MEGA data (adapted from Mark Mikkelsen's Gannet implementation). This can be used as a valid option in case of VAPOR WS which introduces different phases to the transients.

This can be envoked by   opts.SpecReg                = 'ProbSpecReg'; in the job file.