Bug fix un-edited data
Add support for un-edited data with multiple measurements (i.e. extra/subspec dimensions).

The merge of multiple dimensions was done in the process function and did change the raw entries in the prior implementation. Therefore, the GUI crashed when only OspreyLoad was called.

Moved the clean-up in the OspreyLoad function to fix this