Create OspreyHTMLReport.m
Initial commit of Osprey HTML report. This report creates a short subject-specific HTML report to be created after finishing the whole pipeline.

It includes a summary of the alignment process and frequency drift traces. A simplified traffic light system for data quality measures based on Martin Wilson Clinical MRS consensus paper. Summary of the LCM results and fitQA numbers as well as a plot with voxel localization and tissue fractions.

Outputs are BIDS compliant.