[FEATURE REQUEST] - Add external segmentation results - OspreySeg - Helge
Added support for arbitrary external segmentation results to Osprey. The external *.nii/*.nii.gz. results are defined in the jobfile including a 1x3 cell for each subject pointing to the path for the gray matter, white matter, and csf segmentation tissue class maps. The results have to be in the same coordinate space as the supplied anatomical image (file_nii) and are used during the tissue fraction calculations in OspreySeg().

This feature will be used in HBCD.

The example jobfiles were updated accordingly to describe the new feature.