[BUG FIX] - Osprey GUI crash - Model/Coreg button - Lewis Crawford
Fixes issues related to the  GUI crash report on MRSHub (https://forum.mrshub.org/t/error-launching-model-data-and-coregister-in-osprey-gui/859/3).

The crash is potentially caused by the pause on error function in MATLAB, as there were some 'yellow' messages after clicking the model and coreg buttons. Those severe errors which were automatically fixed by MATLAB. However, those would still trigger a pause on error if this options is turned on. This is fixed now.

Thanks for reporting Lewis Crawford