[FEATURE REQUEST] -Manually overwrite subspectra order - OspreyProcess- Helge Zoellner
Added opts field Order to the struct to manually overwrite the sub-spectra order.

For MEGA this is a single number indicating whether the order is switched. Osprey convention is edit OFF first.
For HERMES/HERCULES it is a 4 entry vector indicating the index at which the spectra A,B,C, and D are acquired, e.g.:

GABA-OFF GSH-OFF A
GABA-ON GSH-OFF B
GABA-OFF GSH-ON C
GABA-ON GSH-ON D