[BUG FIX] - Segmentation fails for DCM spec files - OspreyCoreg OspreySeg - Stella Hadfield Elliot
This bug was related to the fact that DCM files have no filename during the fileparts generation of the file input. Therefore, they were always flagged as SameName files. However, without a filename the OspreySeg output resulted in similar file names for all datasets.

This has now been fixed for OspreyCoreg and OspreySeg.