Modify op_lorentzianPeak and op_gaussianPeak for correct area calculation
These changes modify the two functions generation Lorentzian and Gaussian ways such that the amplitude that is entered as the argument corresponds to the *area* under the curve (making the functions much more useful than using the previous amplitude definition). It also performs the 'divide first point by 2' correction (for the first point, FID is only acquired for half the dwell time) that eliminates a small baseline shift that we previously corrected for with op_dccorr.