Adaptions of geometry header from XA to VE in io_loadspec_rda.m
Different nomenclature in XA .rda files is taken into account. If XA30 and the usual fields of the geometry header are empty, other fields will be used and VOIRotationInPlane will be calculated. It has been assumed that SlabOrientation[1] is associatied with Phase vector and SlabOrientation[2] with Readout vector in coreg_siemens.m and the rotation is counterclockwise.