io_loadspec_niimrs.m : bugfix for optional dim_{n}_header
io_loadpsec_niimrs can load additional detail for the "extras" dimension, but did not check for the presence of the corresponding dim_{n}_header before using it. This caused failures when reading data generated by a recent spec2nii.

Furthermore, the dim_number index was not set correctly to the identified "extras" dimension.