Update osp_recalculate_basis_specs.m
Accommodate 'new flavor' of MRSCloud-style basis set simulations.

There have been a few changes in the MRSCloud simulation code that I have been using internally and for collaborators (this will be obsolete in the eventual Osprey 3.0 with the new model. Basically, the metabolite basis functions are shifted in the time domain (by the difference between the slice-selection center frequency and the water resonance) - this resolves a couple not-so-nice effects with the old definitions when the output data were zero-filled. However, this new way of doing things needs to be accommodated here (since Osprey expects .mat basis sets to be in the 'old' convention). Tested so that it does not 'break' the old way of doing things and only triggers when a 'new-flavor' basis set with a 'v2' tag is used.