Update MMLipLCModel.json
Small fix to the definition of MMexp - the `fwmin` and `fwex` values were both set to zero, which led to a funky MM basis function being generated. It was never used anyway (the values in the json are just used to do some amplitude constraining during fitting), but still fixing it here.