Update osp_fitInitialise.m
Fix (finally!) a long-existing bug that occurred when external LCModel-type basis sets had metabolite names that did not match the internal Osprey convention  and were then added to the `chomit` list and therefore excluded from a fit with the LCModel binary (most commonly, e.g., Ins was excluded because it was not named mI).

All metabolite names in external basis sets are now checked against a list of valid abbreviations (listValidBasisFunctioNames) and are *only* excluded if no match is found at all. An info text is given to the console whenever Osprey would prefer a different naming convention.

A similar change will eventually need to be done for the lookup table in OspreyQuantify.