[BUG FIX] Unreasonably slow processing, due to large arbitrary scale on incoming data
op_creChoFit when called by OspreyProcess : on_robustSpecReg : op_measureDrift was unreasonably slow on some datasets, adding some minutes to the OspreyProcess call.

This was due to the lsqcurvefit call running through to the iteration limit, rather than satisfying the convergence criteria (tolerance). This in turn was a result of the arbitrary scale of the inconing data, which can vary by many orders of magnitude between datasets. For datasets with large peak magnitude (eg, in the order of 10^7), tolerance criteria were almost

This also gives very different performance (in terms of runtime, and perhaps modelling precision) depending on an arbitrary scaling factor -- ie, between sites, vendors, etc.

This fix applies a scaling factor to normalise the data before fitting, to ensure consistent performance regardless of scale. Naturally the final result is adjusted back to the original scale. For the tested GE datasets, this reduces processing time by almost two minutes; numeric outcomes indistinguishable.