Cleanup osp_fitInitialise
Suggested fix for #828, #829 (and "better" fix for #798). This patch gives more consistent (and concise!) handling of compiled vs non-compiled modes, more consistent use of specified basissetFolder, and predictable behaviour regardless of whether basissetFile is specified as a single string or a cell array of strings.

Note, basissetFile as a cell might remain problematic for certain fit_*m functions, which are not addressed here.